rgw:Add rgw_dns_name to rgw entity.
rgw must be configured with the hostname and port.

This is due to the rgw being unable to derive them due to implementation
details. This is done with the 'rgw_dns_name' key value pair in the rgw entity.

Without this configuration the 'rgw' cannot work with s3 API as the returned
hostname and port are wrong in the response.

This patch resolves the issue.

Signed-off-by: Owen Synge <osynge@suse.com>